formatted the quickfix class
